Add value field to CBOR to get underlying value
Not sure if there was already a way to do this, but after I decoded a CBOR payload, I was unable to get the underlying value. 

The new value field will return the underlying value as a `Any?` type which gives the developer the ability to cast the value to the type (assuming s/he knows the type).

```swift
if let myString = cbor["myString"]?.value as? String {
    print("My String: \(myString)")
}
```

Let me know what you think.