Fix `Date` encoding and decoding.
Also add support for another encoding and decoding option: `dateStrategy`. This
option accepts a value of type `DateStrategy`. The default value is
`taggedAsEpochTimestamp`, which is what the existing behaviour was. This means
`Date`s are encoded as tagged values, where the tag is `1` and the value is an
epoch timestamp.

The other value possible as a `dateStrategy` is `.annotatedMap`. This will
encode a `Date` as a map that looks like this:

```json
{
  "__type": "date_epoch_timestamp",
  "__value": 123456789
}
```

This is potentially useful for interacting with CBOR implementations that don't
support tagged values.